board_interface: - Implemented ADCs, which are no functional. - Implemented ADC BIT tests. - Data read from interfaces is now in array format, though this may need revisiting. - ADCs are now calibrated based on the calibration diode on the board (that ADC and doide setup is not implmented yet, though)

+++ board_interface/src/Hardware/Device_Pwm_Pca9685.h
@@ -2,7 +2,7 @@
 #define DEVICE_PWM_PCA9685_H
 #include <stdint.h>
 
-const uint16_t MAX_PWM_TICKS = 4095;
-const uint16_t MAX_PWM_FREQUENCY_VALUE = 3500;
+const float MAX_PWM_TICKS = 4095;
+const float MAX_PWM_FREQUENCY_VALUE = 3500;
 
 #endif /* ifndef DEVICE_PWM_PCA9685_H */
